update model, add view and update profile

+++ Services/Services/QuestionServices.cs
@@ -1,8 +1,10 @@
 ﻿using AutoMapper;
+using ExagenSharedProject;
 using Repositories;
 using Repositories.Models;
 using Services.Interfaces;
 using Services.ViewModels;
+using System.Runtime.InteropServices.ComTypes;
 using System.Text.Json;
 
 namespace Services.Services
@@ -19,15 +21,22 @@ namespace Services.Services
         }
 
         public async Task<bool> AddQuestions(QuestionCreate questionCreate, Guid currentUser)
-        {
-            QuestionJson questionJson = new QuestionJson
-            {
-                Content = questionCreate.Content,
-                Answers = questionCreate.Answers,
-            };
-
+        {            
             try
             {
+                // check if answers contains correct answer
+                if (!questionCreate.CorrectAnswers.Any(answer => questionCreate.Answers.Contains(answer)))
+                {
+                    throw new Exception("Câu trả lời đúng không hợp lệ vì không nằm trong các đáp án được đưa ra.");
+                } 
+
+                QuestionJson questionJson = new QuestionJson
+                {
+                    Content = questionCreate.Content,
+                    Answers = questionCreate.Answers,
+                    CorrectAnswers = questionCreate.CorrectAnswers
+                };
+
 
                 if (questionCreate.SectionId != null)
                 {
@@ -47,18 +56,22 @@ namespace Services.Services
                     }
                 }
 
-                if (questionCreate.IsUseToSell < 1 || questionCreate.IsUseToSell > 2)
+                if (questionCreate.Subject != null && (questionCreate.Subject < 1 || questionCreate.Subject > 3))
                 {
-                    throw new Exception("Trạng thái không hợp lệ");
+                    throw new Exception("Môn học không hợp lệ");
                 }
 
                 var question = _mapper.Map<Question>(questionCreate);
-                question.QuestionContent = JsonSerializer.Serialize((questionJson));
+                question.QuestionContent = JsonSerializer.Serialize(questionJson, new JsonSerializerOptions
+                {
+                    WriteIndented = true,
+                    Encoder = System.Text.Encodings.Web.JavaScriptEncoder.UnsafeRelaxedJsonEscaping
+                });
                 question.CreatedOn = DateTime.Now;
                 question.ModifiedOn = DateTime.Now;
                 question.CreatedBy = currentUser;
                 question.ModifiedBy = currentUser;
-                question.Status = questionCreate.IsUseToSell;
+                question.Status = 0;
 
                 _unitOfWork.QuestionRepo.AddAsync(question);
                 var result = await _unitOfWork.SaveChangesAsync();
@@ -76,14 +89,23 @@ namespace Services.Services
             }
         }
 
-        public async Task<bool> DeleteQuestion(Guid questionId, Guid currentUser)
+        public async Task<bool> DeleteQuestion(Guid questionId, Guid currentUserId, int grade)
         {
             try
             {
-                var question = await _unitOfWork.QuestionRepo.FindByField(question => question.QuestionId == questionId);
+                var question = await _unitOfWork.QuestionRepo.FindByField(question => question.QuestionId == questionId && question.Grade == grade);
                 if (question == null)
                 {
                     throw new Exception("Câu hỏi không tồn tại.");
+                }
+
+                var currentUser = await _unitOfWork.UserRepo.FindByField(user => user.UserId == currentUserId);
+
+                var isShare = await CheckPermissionForQuestion(questionId, 3, currentUserId, currentUser.SchoolId);
+
+                if (question.CreatedBy != currentUserId &&  !isShare)
+                {
+                    throw new Exception("Bạn không có quyền xóa câu hỏi này.");
                 }
 
                 _unitOfWork.QuestionRepo.Remove(question);
@@ -103,11 +125,11 @@ namespace Services.Services
             }
         }
 
-        public async Task<IEnumerable<QuestionViewModels>> GetAllQuestion()
+        public async Task<IEnumerable<QuestionViewModels>> GetAllMyQuestionByGrade(int grade, Guid currentUserId)
         {
             try
             {
-                var questions = await _unitOfWork.QuestionRepo.GetAllAsync();
+                var questions = await _unitOfWork.QuestionRepo.FindListByField(questions => questions.Grade == grade && questions.CreatedBy == currentUserId);
                 if (questions == null)
                 {
                     return null;
@@ -122,14 +144,46 @@ namespace Services.Services
             }
         }
 
-        public async Task<QuestionViewModels> GetQuestionByQuestionId(Guid questionId)
+        public async Task<IEnumerable<QuestionViewModels>> GetAllValidQuestionByGradeForMe(int grade, Guid currentUserId)
         {
             try
             {
-                var question = await _unitOfWork.QuestionRepo.FindByField(question => question.QuestionId == questionId);
+                var currentUser = await _unitOfWork.UserRepo.FindByField(user => user.UserId == currentUserId);
+
+                var listQuestionId = await GetQuestionIdIsSharedToRead(currentUserId, currentUser.SchoolId);
+
+                var questions = await _unitOfWork.QuestionRepo.FindListByField(question => question.Grade == grade && 
+                                                                                            (listQuestionId.Contains(question.QuestionId) || question.CreatedBy == currentUserId));
+
+                if (questions == null)
+                {
+                    return null;
+                }
+
+                var questionViewModels = _mapper.Map<IEnumerable<QuestionViewModels>>(questions);
+                return questionViewModels;
+
+            }
+            catch (Exception e)
+            {
+                throw new Exception("Lỗi ở QuestionServices - GetAllValidQuestionByGradeForMe: " + e.Message);
+            }
+        }
+
+        public async Task<QuestionViewModels> GetQuestionByQuestionIdAndGrade(Guid questionId, int grade, Guid currentUserId)
+        {
+            try
+            {
+                var question = await _unitOfWork.QuestionRepo.FindByField(question => question.Grade == grade && question.QuestionId == questionId);
                 if (question == null)
                 {
                     throw new Exception("Câu hỏi không tồn tại.");
+                }
+
+                var currentUser = await _unitOfWork.UserRepo.FindByField(user => user.UserId == currentUserId);
+                if (!(await CheckPermissionForQuestion(questionId, 0, currentUserId, currentUser.SchoolId)))
+                {
+                    throw new Exception("Bạn không có quyền xem câu hỏi này.");
                 }
 
                 var questionViewModels = _mapper.Map<QuestionViewModels>(question);
@@ -141,18 +195,62 @@ namespace Services.Services
             }
         }
 
-        public async Task<bool> UpdateQuestion(QuestionUpdate questionUpdate, Guid currentUser)
+        public async Task<IEnumerable<QuestionViewModels>> GetQuestionBySectionIdAndGrade(Guid sectionId, int grade)
         {
-
-            QuestionJson questionJson = new QuestionJson
+            try
             {
-                Content = questionUpdate.Content,
-                Answers = questionUpdate.Answers,
-            };
+                var questions = await _unitOfWork.QuestionRepo.FindListByField(question => question.SectionId == sectionId && question.Grade == grade);
+                if (questions == null)
+                {
+                    return null;
+                }
+
+                var questionViewModels = _mapper.Map<IEnumerable<QuestionViewModels>>(questions);
+                return questionViewModels;
+            }
+            catch (Exception e)
+            {
+                throw new Exception("Lỗi ở QuestionServices - GetQuestionBySectionId: " + e.Message);
+            }
+        }
+
+        public async Task<IEnumerable<QuestionViewModels>> GetQuestionBySubjectAndGrade(int subject, int grade)
+        {
+            try
+            {
+                var questions = await _unitOfWork.QuestionRepo.FindListByField(question => question.Subject == subject && question.Grade == grade);
+                if (questions == null)
+                {
+                    return null;
+                }
+
+                var questionViewModels = _mapper.Map<IEnumerable<QuestionViewModels>>(questions);
+                return questionViewModels;
+            }
+            catch (Exception e)
+            {
+                throw new Exception("Lỗi ở QuestionServices - GetQuestionBySubject: " + e.Message);
+            }
+        }
+
+        public async Task<bool> UpdateQuestion(QuestionUpdate questionUpdate, Guid currentUser, int grade)
+        {
 
             try
             {
-                var question = await _unitOfWork.QuestionRepo.FindByField(question => question.QuestionId == questionUpdate.QuestionId);
+                if (!questionUpdate.CorrectAnswers.Any(answer => questionUpdate.Answers.Contains(answer)))
+                {
+                    throw new Exception("Câu trả lời đúng không hợp lệ vì không nằm trong các đáp án được đưa ra.");
+                }
+
+                QuestionJson questionJson = new QuestionJson
+                {
+                    Content = questionUpdate.Content,
+                    Answers = questionUpdate.Answers,
+                    CorrectAnswers = questionUpdate.CorrectAnswers
+                };
+
+                var question = await _unitOfWork.QuestionRepo.FindByField(question => question.QuestionId == questionUpdate.QuestionId && question.Grade == grade);
                 if (question == null)
                 {
                     throw new Exception("Câu hỏi không tồn tại.");
@@ -176,16 +274,11 @@ namespace Services.Services
                     }
                 }
 
-                if (questionUpdate.IsUseToSell < 1 || questionUpdate.IsUseToSell > 2)
-                {
-                    throw new Exception("Trạng thái không hợp lệ");
-                }
 
                 question = _mapper.Map(questionUpdate, question);
                 question.QuestionContent = JsonSerializer.Serialize((questionJson));
                 question.ModifiedOn = DateTime.Now;
                 question.ModifiedBy = currentUser;
-                question.Status = questionUpdate.IsUseToSell;
 
                 _unitOfWork.QuestionRepo.AddAsync(question);
                 var result = await _unitOfWork.SaveChangesAsync();
@@ -199,6 +292,71 @@ namespace Services.Services
             catch (Exception e)
             {
                 throw new Exception("Lỗi ở QuestionServices - UpdateQuestion: " + e.Message);
+            }
+        }
+
+        private async Task<bool> CheckPermissionForQuestion(Guid questionId, int permissionType, Guid userId, Guid? schoolId)
+        {
+            if (schoolId != null)
+            {
+                var share = await _unitOfWork.ShareRepo.FindByField(share => share.QuestionId == questionId && share.PermissionType >= permissionType &&
+                                                                                share.UserId == userId);
+                if (share == null)
+                {
+                    return false;
+                }
+
+                return true;
+
+            }
+            else
+            {
+                var share = await _unitOfWork.ShareRepo.FindByField(share => share.QuestionId == questionId && share.PermissionType >= permissionType &&
+                                                                                (share.UserId == userId || share.SchoolId == schoolId));
+                if (share == null)
+                {
+                    return false;
+                }
+
+                return true;
+            }
+        }
+    
+        private async Task<IEnumerable<Guid>> GetQuestionIdIsSharedToRead(Guid userId, Guid? schoolId)
+        {
+            if (schoolId != null)
+            {
+                var shares = await _unitOfWork.ShareRepo.FindListByField(share => share.UserId == userId);
+                if (shares == null)
+                {
+                    return null;
+                }
+
+                var questionIds = shares.Select(share => share.QuestionId);
+                
+                if (questionIds == null)
+                {
+                    return null;
+                }
+
+                return (IEnumerable<Guid>)questionIds;
+            }
+            else
+            {
+                var shares = await _unitOfWork.ShareRepo.FindListByField(share => share.SchoolId == schoolId);
+                if (shares == null)
+                {
+                    return null;
+                }
+
+                var questionIds = shares.Select(share => share.QuestionId);
+
+                if (questionIds == null)
+                {
+                    return null;
+                }
+
+                return (IEnumerable<Guid>)questionIds;
             }
         }
     }
